Fixbug #230
